requirements.txt
autopep8==1.5.7
numpy==1.19.5
packaging==20.9
pandas==1.2.4
pytest==6.2.4
PyYAML==5.4.1
snorkel==0.9.7
tqdm==4.60.0
twint @ git+https://github.com/twintproject/twint.git@e7c8a0c764f6879188e5c21e25fb6f1f856a7221
